Backend/requirements.txt
flask==3.1.2
gunicorn==23.0.0
pillow==11.3.0
flask-cors==4.0.0
